+++ src/View/Cell/PhotosCell.php
@@ -1,30 +1,30 @@
 <?php
 /**
- * This file is part of MeInstagram.
+ * This file is part of me-cms-instagram.
  *
- * MeInstagram is free software: you can redistribute it and/or modify
+ * me-cms-instagram is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
- * MeInstagram is distributed in the hope that it will be useful,
+ * me-cms-instagram is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with MeInstagram.  If not, see <http://www.gnu.org/licenses/>.
+ * along with me-cms-instagram.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-namespace MeInstagram\View\Cell;
+namespace MeCmsInstagram\View\Cell;
 
 use Cake\Cache\Cache;
 use Cake\View\Cell;
-use MeInstagram\Utility\Instagram;
+use MeCmsInstagram\Utility\Instagram;
 
 /**
  * Photos cell
@@ -32,10 +32,20 @@ use MeInstagram\Utility\Instagram;
 class PhotosCell extends Cell
 {
     /**
+     * Returns an `Instagram` instance
+     * @return \MeCmsInstagram\Utility\Instagram
+     */
+    protected function _getInstagramInstance()
+    {
+        return new Instagram;
+    }
+
+    /**
      * Internal method to get the latest photos
      * @param int $limit Limit
      * @return array
-     * @uses MeInstagram\Utility\Instagram::recent()
+     * @uses MeCmsInstagram\Utility\Instagram::recent()
+     * @uses _getInstagramInstance()
      */
     protected function _latest($limit = 15)
     {
@@ -44,7 +54,7 @@ class PhotosCell extends Cell
 
         //If the data are not available from the cache
         if (empty($photos)) {
-            list($photos) = Instagram::recent(null, $limit);
+            list($photos) = $this->_getInstagramInstance()->recent(null, $limit);
 
             Cache::write($cache, $photos, 'instagram');
         }