+++ tests/TestCase/Utility/SitemapTest.php
@@ -20,24 +20,28 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
+namespace MeCmsInstagram\Test\TestCase\Utility;
 
-if (empty($photos)) {
-    return;
-}
+use Cake\Routing\Router;
+use Cake\TestSuite\TestCase;
+use MeCmsInstagram\Utility\Sitemap;
 
-$this->extend('MeCms./Common/widget');
-$this->assign('title', __dn(
-    'me_cms',
-    'Random photo',
-    'Random {0} photos',
-    count($photos),
-    count($photos)
-));
-
-foreach ($photos as $photo) {
-    echo $this->Html->link(
-        $this->Thumb->fit($photo->path, ['width' => 253]),
-        ['_name' => 'instagramPhotos'],
-        ['class' => 'thumbnail', 'title' => $photo->description]
-    );
+/**
+ * SitemapTest class
+ */
+class SitemapTest extends TestCase
+{
+    /**
+     * Test for `instagram()` method
+     * @test
+     */
+    public function testInstagram()
+    {
+        $this->assertEquals([
+            [
+                'loc' => Router::url(['_name' => 'instagramPhotos'], true),
+                'priority' => '0.5',
+            ],
+        ], Sitemap::instagram());
+    }
 }