+++ tests/test_app/TestApp/config/me_cms_instagram.php
@@ -1,0 +1,7 @@
+<?php
+return [
+    'Instagram' => [
+        //API access token
+        'key' => '00000000',
+    ],
+];